Move source files into logical packages
