+++ src/main/java/bezdelnik/Main.java
@@ -2,11 +2,14 @@ package bezdelnik;
 
 import java.io.IOException;
 
+import bezdelnik.utils.Bezdelnik;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+
+
 
 /**
  * A GUI for Bezdelnik using FXML.
