+++ src/main/java/bezdelnik/utils/BezdelnikException.java
@@ -1,4 +1,4 @@
-package bezdelnik;
+package bezdelnik.utils;
 
 /**
  * Exception specific to the Bezdelnik application.
